[refactor] Provide a gateway to call CandyShop backend APIs
[LIQ-537]

Currently, some information based APIs from backend must need a
CandyShop instance to call with which is not necessary, it lack of
flexibility to be adopted in difference cases.

CandyShop instance should be a part of sdk and not a major character,
the jobs only belongs to it are the transaction methods.

Separate the backend APIs into CandyShopInfoAPI.ts and not modify
any existed public method to keep backward compatibility.