[sdk] Update configBaseUrl calling order
To initiate the EVM CandyShop, SDK will first need to know
the backend to call to fetch the correct shop object. Therefore
the call to configBaseUrl will preceed the new EthCandyShop function.

Since EthCandyShop extends BaseShop and BaseShop calls configBaseUrl
as well in this constructor, there will be 2 configBaseCall calls
for EthCandyShop for time being.